cleanup planning and priorities
